+++ planets-fact-site/app/[planet]/page.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { PlanetInfoPage } from "@components";
+
+export default function Home() {
+  return <PlanetInfoPage />;
+}
