+++ planets-fact-site/app/components/index.ts
@@ -1,2 +1,3 @@
 export { default as CardInformation } from "./card-information";
 export { default as Header } from "./header";
+export { default as PlanetInfoPage } from "./planet-info-page";
